divide ny-page into internal and external
